chore: de duplicating code
