+++ packages/amplify-migration-tests/src/migration-helpers/check-version.ts
@@ -15,6 +15,19 @@ export function versionCheck(cwd: string, testingWithLatestCodebase = false, ver
 }
 
 /**
+ * Validates from and to versions for migration tests.
+ */
+export const validateVersionsForMigrationTest = async (): Promise<void> => {
+  const migrateFromVersion = { v: 'unintialized' };
+  const migrateToVersion = { v: 'unintialized' };
+  await versionCheck(process.cwd(), false, migrateFromVersion);
+  await versionCheck(process.cwd(), true, migrateToVersion);
+  console.log(`Test migration from: ${migrateFromVersion.v} to ${migrateToVersion.v}`);
+  expect(migrateFromVersion.v).not.toEqual(migrateToVersion.v);
+  expect(allowedVersionsToMigrateFrom).toContain(migrateFromVersion.v);
+};
+
+/**
  * This list is used to check migration tests with the following changes. (excludes layer migration tests)
  *
  * api add/update flow: https://github.com/aws-amplify/amplify-cli/pull/8287
